#105 Fix 8.4 and 8.5 deprecations, mostly in Admin Page Framework.
Version 1.5.7, WordPress 7.0

+++ code/health.php
@@ -233,7 +233,7 @@ class Health {
         $max = - $min;
         foreach ( $splits as $split ) {
           if ( is_numeric( $split ) ) {
-            $split = (double) $split;
+            $split = (float) $split;
             $min   = min( $min, $split );
             $max   = max( $max, $split );
           }
